refactor(workspace, cli): unify action-tree walks behind describeActions
Promote the action-tree → ActionManifest walker to a public
`describeActions(actions)` in `@epicenter/workspace`, paired with the
existing `system.describe` RPC and `peerSystem(...).describe()` proxy
(same verb, local vs wire). Collapses two duplicates:

- `attach-sync.ts`'s private `collectActionManifest`/`walkActions`
- `cli/util/local-manifest.ts`'s `collectLocalManifest` (file deleted)

Also replaces `run.ts`'s private `findAction` with the existing public
`resolveActionPath` — they were byte-identical apart from a parameter
type. Renames the CLI-local helper `describeAction` (singular) to
`toActionDescriptor` to avoid sibling-name confusion with the new export.

Tightens the `actions.ts` module header to a visual block showing the
`Actions ↔ ActionManifest` duality; the prior 50-line prose said
"the callable IS the handler" five different ways.

No behavior change. 638 tests pass.
